Factor out cli_options to separate file
